+++ packages/konsole-app/src/components/RecordTypeSelector.tsx
@@ -2,10 +2,9 @@
 // Copyright 2020 DXOS.org
 //
 
-import clsx from 'clsx';
 import React from 'react';
 
-import { makeStyles, Button, ButtonGroup } from '@material-ui/core';
+import { Button, ButtonGroup } from '@material-ui/core';
 
 import { IRecordType } from '../hooks';
 
@@ -15,28 +14,19 @@ interface RecordTypeSelectorProperties {
   onTypeChange: (type: string) => void
 }
 
-const useStyles = makeStyles(theme => ({
-  selected: {
-    color: theme.palette.text.primary
-  }
-}));
-
 export const RecordTypeSelector = ({ types, type: selected, onTypeChange }: RecordTypeSelectorProperties) => {
-  const classes = useStyles();
-
   return (
     <ButtonGroup
       disableRipple
       disableFocusRipple
       variant='outlined'
-      color='primary'
       size='small'
       aria-label='text primary button group'
     >
       {types.map(({ type, label }) => (
         <Button
           key={type}
-          className={clsx({ [classes.selected]: type === selected })}
+          variant={type === selected ? 'contained' : 'outlined'}
           onClick={() => onTypeChange(type)}
         >
           {label}
