+++ packages/konsole-app/src/main.tsx
@@ -13,11 +13,10 @@ import {
   useParams
 } from 'react-router-dom';
 
-import { Container, IPanel, Root, Sidebar } from './containers';
+import { Container, Root, Sidebar } from './containers';
 import { ConfigContext, IConfig, RegistryContext } from './hooks';
-import { ConfigPanel, RecordPanel } from './panels';
+import { panels } from './panels';
 import { MockRegistryClient } from './testing';
-import { Config } from '../stories/panels.stories';
 
 // TODO(burdon): Load from environment.
 const config: IConfig = {
@@ -26,19 +25,6 @@ const config: IConfig = {
     theme: 'dark'
   }
 };
-
-const panels: IPanel[] = [
-  {
-    path: '/config',
-    label: 'Config',
-    component: ConfigPanel
-  },
-  {
-    path: '/records',
-    label: 'Records',
-    component: RecordPanel
-  }
-];
 
 const App = () => {
   const history = useHistory();
